Avoid webworker types in @iov/jsonrpc types
The types Worker, MessageEvent, Transferable broke import of
@iov/jsonrpc in builds that do not have webworker
