+++ packages/iov-jsonrpc/types/index.d.ts
@@ -1,4 +1,4 @@
-export { JsonRpcClient } from "./jsonrpcclient";
+export { JsonRpcClient, SimpleMessagingConnection } from "./jsonrpcclient";
 export { JsonCompatibleArray, JsonCompatibleDictionary, JsonCompatibleValue, isJsonCompatibleArray, isJsonCompatibleDictionary, isJsonCompatibleValue, } from "./jsoncompatibledictionary";
 export { parseJsonRpcId, parseJsonRpcRequest, parseJsonRpcResponse, parseJsonRpcError } from "./parse";
 export { isJsonRpcErrorResponse, JsonRpcRequest, JsonRpcErrorResponse, JsonRpcResponse, jsonRpcCodeInternalError, jsonRpcCodeInvalidParams, jsonRpcCodeInvalidRequest, jsonRpcCodeMethodNotFound, jsonRpcCodeParseError, jsonRpcCodeServerErrorDefault, } from "./types";
