+++ packages/iov-jsonrpc/types/jsonrpcclient.d.ts
@@ -1,12 +1,11 @@
-import { JsonRpcRequest, JsonRpcResponse } from "./types";
-interface SimpleMessagingConnection {
-    onmessage: ((this: Worker, ev: MessageEvent) => any) | null;
-    readonly postMessage: (message: any, transfer?: Transferable[]) => void;
+import { Stream } from "xstream";
+import { JsonRpcErrorResponse, JsonRpcRequest, JsonRpcResponse } from "./types";
+export interface SimpleMessagingConnection {
+    readonly responseStream: Stream<JsonRpcResponse | JsonRpcErrorResponse>;
+    readonly sendRequest: (request: JsonRpcRequest) => void;
 }
 export declare class JsonRpcClient {
-    private readonly responseStream;
     private readonly connection;
     constructor(connection: SimpleMessagingConnection);
     run(request: JsonRpcRequest): Promise<JsonRpcResponse>;
 }
-export {};
